requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
sqlalchemy==2.0.46
pydantic==2.12.5
alembic==1.18.3
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
boto3==1.34.0
